+++ tests/conftest.py
@@ -1,3 +1,5 @@
+import os
+
 import pytest
 from utils import assert_files_equal, datapath, cutpath
 from cutadapt.__main__ import main
@@ -14,6 +16,7 @@ def run(tmp_path):
     def _run(params, expected, inpath) -> Statistics:
         if type(params) is str:
             params = params.split()
+        params += ["--json", os.fspath(tmp_path / "stats.cutadapt.json")]
         tmp_fastaq = tmp_path / expected
         params += ['-o', tmp_fastaq]
         params += [datapath(inpath)]